+++ pkg/querier/tripperware/queryrange/results_cache_test.go
@@ -9,6 +9,7 @@ import (
 	"time"
 
 	"github.com/go-kit/log"
+	"github.com/gogo/protobuf/proto"
 	"github.com/gogo/protobuf/types"
 	"github.com/prometheus/common/model"
 	"github.com/stretchr/testify/assert"
@@ -18,6 +19,7 @@ import (
 	"github.com/cortexproject/cortex/pkg/chunk/cache"
 	"github.com/cortexproject/cortex/pkg/cortexpb"
 	"github.com/cortexproject/cortex/pkg/querier/tripperware"
+	"github.com/cortexproject/cortex/pkg/tenant"
 	"github.com/cortexproject/cortex/pkg/util/flagext"
 )
 
@@ -119,11 +121,11 @@ var (
 	}
 )
 
-func mkAPIResponse(start, end, step int64) *tripperware.PrometheusResponse {
-	return mkAPIResponseWithStats(start, end, step, false)
+func mkAPIResponse(start, end, step int64) tripperware.Response {
+	return mkAPIResponseWithStats(start, end, step, false, false)
 }
 
-func mkAPIResponseWithStats(start, end, step int64, withStats bool) *tripperware.PrometheusResponse {
+func mkAPIResponseWithStats(start, end, step int64, withStats bool, oldFormat bool) tripperware.Response {
 	var samples []cortexpb.Sample
 	var stats *tripperware.PrometheusResponseStats
 	if withStats {
@@ -142,6 +144,24 @@ func mkAPIResponseWithStats(start, end, step int64, withStats bool) *tripperware
 			})
 
 			stats.Samples.TotalQueryableSamples += i
+		}
+	}
+
+	if oldFormat {
+		return &PrometheusResponse{
+			Status: StatusSuccess,
+			Data: PrometheusData{
+				ResultType: matrix,
+				Stats:      stats,
+				Result: []tripperware.SampleStream{
+					{
+						Labels: []cortexpb.LabelAdapter{
+							{Name: "foo", Value: "bar"},
+						},
+						Samples: samples,
+					},
+				},
+			},
 		}
 	}
 
@@ -169,19 +189,27 @@ func mkAPIResponseWithStats(start, end, step int64, withStats bool) *tripperware
 }
 
 func mkExtentWithStats(start, end int64) tripperware.Extent {
-	return mkExtentWithStepWithStats(start, end, 10, true)
+	return mkExtentWithStepWithStatsWithFormat(start, end, 10, true, false)
 }
 
 func mkExtent(start, end int64) tripperware.Extent {
-	return mkExtentWithStepWithStats(start, end, 10, false)
+	return mkExtentWithStepWithStatsWithFormat(start, end, 10, false, false)
+}
+
+func mkExtentWithOldFormat(start, end int64) tripperware.Extent {
+	return mkExtentWithStepWithStatsWithFormat(start, end, 10, false, true)
 }
 
 func mkExtentWithStep(start, end, step int64) tripperware.Extent {
-	return mkExtentWithStepWithStats(start, end, step, false)
+	return mkExtentWithStepWithStatsWithFormat(start, end, step, false, false)
 }
 
-func mkExtentWithStepWithStats(start, end, step int64, withStats bool) tripperware.Extent {
-	res := mkAPIResponseWithStats(start, end, step, withStats)
+func mkExtentWithStepAndFormat(start, end, step int64, oldFormat bool) tripperware.Extent {
+	return mkExtentWithStepWithStatsWithFormat(start, end, step, false, false)
+}
+
+func mkExtentWithStepWithStatsWithFormat(start, end, step int64, withStats bool, oldFormat bool) tripperware.Extent {
+	res := mkAPIResponseWithStats(start, end, step, withStats, oldFormat)
 	any, err := types.MarshalAny(res)
 	if err != nil {
 		panic(err)
@@ -214,29 +242,29 @@ func TestStatsCacheQuerySamples(t *testing.T) {
 			name:                       "should return response with stats",
 			cacheQueryableSamplesStats: true,
 			req:                        noCacheRequestWithStats,
-			upstreamResponse:           mkAPIResponseWithStats(0, 100, 10, true),
-			expectedResponse:           mkAPIResponseWithStats(0, 100, 10, true),
+			upstreamResponse:           mkAPIResponseWithStats(0, 100, 10, true, false),
+			expectedResponse:           mkAPIResponseWithStats(0, 100, 10, true, false),
 		},
 		{
 			name:                       "should return response strip stats if not requested",
 			cacheQueryableSamplesStats: true,
 			req:                        noCacheRequest,
-			upstreamResponse:           mkAPIResponseWithStats(0, 100, 10, false),
-			expectedResponse:           mkAPIResponseWithStats(0, 100, 10, false),
+			upstreamResponse:           mkAPIResponseWithStats(0, 100, 10, false, false),
+			expectedResponse:           mkAPIResponseWithStats(0, 100, 10, false, false),
 		},
 		{
 			name:                       "should not ask stats is cacheQueryableSamplesStats is disabled",
 			cacheQueryableSamplesStats: false,
 			req:                        noCacheRequest,
-			upstreamResponse:           mkAPIResponseWithStats(0, 100, 10, false),
-			expectedResponse:           mkAPIResponseWithStats(0, 100, 10, false),
+			upstreamResponse:           mkAPIResponseWithStats(0, 100, 10, false, false),
+			expectedResponse:           mkAPIResponseWithStats(0, 100, 10, false, false),
 		},
 		{
 			name:                       "should not forward stats when cacheQueryableSamplesStats is disabled",
 			cacheQueryableSamplesStats: false,
 			req:                        noCacheRequestWithStats,
-			upstreamResponse:           mkAPIResponseWithStats(0, 100, 10, true),
-			expectedResponse:           mkAPIResponseWithStats(0, 100, 10, false),
+			upstreamResponse:           mkAPIResponseWithStats(0, 100, 10, true, false),
+			expectedResponse:           mkAPIResponseWithStats(0, 100, 10, false, false),
 		},
 	} {
 		tc := tc
@@ -533,6 +561,20 @@ func TestPartition(t *testing.T) {
 		},
 
 		{
+			name: "Test a complete hit with old format.",
+			input: &tripperware.PrometheusRequest{
+				Start: 0,
+				End:   100,
+			},
+			prevCachedResponse: []tripperware.Extent{
+				mkExtentWithOldFormat(0, 100),
+			},
+			expectedCachedResponse: []tripperware.Response{
+				mkAPIResponse(0, 100, 10),
+			},
+		},
+
+		{
 			name: "Test with a complete miss.",
 			input: &tripperware.PrometheusRequest{
 				Start: 0,
@@ -567,6 +609,25 @@ func TestPartition(t *testing.T) {
 			},
 		},
 		{
+			name: "Test a partial hit with old format.",
+			input: &tripperware.PrometheusRequest{
+				Start: 0,
+				End:   100,
+			},
+			prevCachedResponse: []tripperware.Extent{
+				mkExtentWithOldFormat(50, 100),
+			},
+			expectedRequests: []tripperware.Request{
+				&tripperware.PrometheusRequest{
+					Start: 0,
+					End:   50,
+				},
+			},
+			expectedCachedResponse: []tripperware.Response{
+				mkAPIResponse(50, 100, 10),
+			},
+		},
+		{
 			name: "Test multiple partial hits.",
 			input: &tripperware.PrometheusRequest{
 				Start: 100,
@@ -588,6 +649,27 @@ func TestPartition(t *testing.T) {
 			},
 		},
 		{
+			name: "Test multiple partial hits with old format.",
+			input: &tripperware.PrometheusRequest{
+				Start: 100,
+				End:   200,
+			},
+			prevCachedResponse: []tripperware.Extent{
+				mkExtentWithOldFormat(50, 120),
+				mkExtentWithOldFormat(160, 250),
+			},
+			expectedRequests: []tripperware.Request{
+				&tripperware.PrometheusRequest{
+					Start: 120,
+					End:   160,
+				},
+			},
+			expectedCachedResponse: []tripperware.Response{
+				mkAPIResponse(100, 120, 10),
+				mkAPIResponse(160, 200, 10),
+			},
+		},
+		{
 			name: "Partial hits with tiny gap.",
 			input: &tripperware.PrometheusRequest{
 				Start: 100,
@@ -596,6 +678,26 @@ func TestPartition(t *testing.T) {
 			prevCachedResponse: []tripperware.Extent{
 				mkExtent(50, 120),
 				mkExtent(122, 130),
+			},
+			expectedRequests: []tripperware.Request{
+				&tripperware.PrometheusRequest{
+					Start: 120,
+					End:   160,
+				},
+			},
+			expectedCachedResponse: []tripperware.Response{
+				mkAPIResponse(100, 120, 10),
+			},
+		},
+		{
+			name: "Partial hits with tiny gap with old format.",
+			input: &tripperware.PrometheusRequest{
+				Start: 100,
+				End:   160,
+			},
+			prevCachedResponse: []tripperware.Extent{
+				mkExtentWithOldFormat(50, 120),
+				mkExtentWithOldFormat(122, 130),
 			},
 			expectedRequests: []tripperware.Request{
 				&tripperware.PrometheusRequest{
@@ -638,6 +740,20 @@ func TestPartition(t *testing.T) {
 			},
 		},
 		{
+			name: "Test when hit has a large step and only a single sample extent with old format.",
+			// If there is a only a single sample in the split interval, start and end will be the same.
+			input: &tripperware.PrometheusRequest{
+				Start: 100,
+				End:   100,
+			},
+			prevCachedResponse: []tripperware.Extent{
+				mkExtentWithOldFormat(100, 100),
+			},
+			expectedCachedResponse: []tripperware.Response{
+				mkAPIResponse(100, 105, 10),
+			},
+		},
+		{
 			name: "[Stats] Test a complete hit.",
 			input: &tripperware.PrometheusRequest{
 				Start: 0,
@@ -647,7 +763,21 @@ func TestPartition(t *testing.T) {
 				mkExtentWithStats(0, 100),
 			},
 			expectedCachedResponse: []tripperware.Response{
-				mkAPIResponseWithStats(0, 100, 10, true),
+				mkAPIResponseWithStats(0, 100, 10, true, false),
+			},
+		},
+
+		{
+			name: "[Stats] Test a complete hit with old format.",
+			input: &tripperware.PrometheusRequest{
+				Start: 0,
+				End:   100,
+			},
+			prevCachedResponse: []tripperware.Extent{
+				mkExtentWithStepWithStatsWithFormat(0, 100, 10, true, true),
+			},
+			expectedCachedResponse: []tripperware.Response{
+				mkAPIResponseWithStats(0, 100, 10, true, false),
 			},
 		},
 
@@ -659,6 +789,21 @@ func TestPartition(t *testing.T) {
 			},
 			prevCachedResponse: []tripperware.Extent{
 				mkExtentWithStats(110, 210),
+			},
+			expectedRequests: []tripperware.Request{
+				&tripperware.PrometheusRequest{
+					Start: 0,
+					End:   100,
+				}},
+		},
+		{
+			name: "[Stats] Test with a complete miss with old format.",
+			input: &tripperware.PrometheusRequest{
+				Start: 0,
+				End:   100,
+			},
+			prevCachedResponse: []tripperware.Extent{
+				mkExtentWithStepWithStatsWithFormat(110, 210, 10, true, true),
 			},
 			expectedRequests: []tripperware.Request{
 				&tripperware.PrometheusRequest{
@@ -682,7 +827,7 @@ func TestPartition(t *testing.T) {
 				},
 			},
 			expectedCachedResponse: []tripperware.Response{
-				mkAPIResponseWithStats(50, 100, 10, true),
+				mkAPIResponseWithStats(50, 100, 10, true, false),
 			},
 		},
 		{
@@ -702,8 +847,8 @@ func TestPartition(t *testing.T) {
 				},
 			},
 			expectedCachedResponse: []tripperware.Response{
-				mkAPIResponseWithStats(100, 120, 10, true),
-				mkAPIResponseWithStats(160, 200, 10, true),
+				mkAPIResponseWithStats(100, 120, 10, true, false),
+				mkAPIResponseWithStats(160, 200, 10, true, false),
 			},
 		},
 		{
@@ -723,7 +868,7 @@ func TestPartition(t *testing.T) {
 				},
 			},
 			expectedCachedResponse: []tripperware.Response{
-				mkAPIResponseWithStats(100, 120, 10, true),
+				mkAPIResponseWithStats(100, 120, 10, true, false),
 			},
 		},
 		{
@@ -753,7 +898,21 @@ func TestPartition(t *testing.T) {
 				mkExtentWithStats(100, 100),
 			},
 			expectedCachedResponse: []tripperware.Response{
-				mkAPIResponseWithStats(100, 105, 10, true),
+				mkAPIResponseWithStats(100, 105, 10, true, false),
+			},
+		},
+		{
+			name: "[stats] Test when hit has a large step and only a single sample extent with old format.",
+			// If there is a only a single sample in the split interval, start and end will be the same.
+			input: &tripperware.PrometheusRequest{
+				Start: 100,
+				End:   100,
+			},
+			prevCachedResponse: []tripperware.Extent{
+				mkExtentWithStepWithStatsWithFormat(100, 100, 10, true, true),
+			},
+			expectedCachedResponse: []tripperware.Response{
+				mkAPIResponseWithStats(100, 105, 10, true, false),
 			},
 		},
 	} {
@@ -958,6 +1117,54 @@ func TestHandleHit(t *testing.T) {
 				mkExtentWithStep(60, 180, 20),
 			},
 		},
+		{
+			name: "Should replace tiny extents with old format that are cover by bigger request",
+			input: &tripperware.PrometheusRequest{
+				Start: 100,
+				End:   200,
+				Step:  5,
+			},
+			cachedEntry: []tripperware.Extent{
+				mkExtentWithStepAndFormat(0, 50, 5, true),
+				mkExtentWithStepAndFormat(60, 65, 5, true),
+				mkExtentWithStepAndFormat(100, 105, 5, true),
+				mkExtentWithStepAndFormat(110, 115, 5, true),
+				mkExtentWithStepAndFormat(120, 125, 5, true),
+				mkExtentWithStepAndFormat(220, 225, 5, true),
+				mkExtentWithStepAndFormat(240, 250, 5, true),
+			},
+			expectedUpdatedCachedEntry: []tripperware.Extent{
+				mkExtentWithStep(0, 50, 5),
+				mkExtentWithStep(60, 65, 5),
+				mkExtentWithStep(100, 200, 5),
+				mkExtentWithStep(220, 225, 5),
+				mkExtentWithStep(240, 250, 5),
+			},
+		},
+		{
+			name: "Should replace tiny extents with mixed of new and old format that are cover by bigger request",
+			input: &tripperware.PrometheusRequest{
+				Start: 100,
+				End:   200,
+				Step:  5,
+			},
+			cachedEntry: []tripperware.Extent{
+				mkExtentWithStepAndFormat(0, 50, 5, true),
+				mkExtentWithStepAndFormat(60, 65, 5, false),
+				mkExtentWithStepAndFormat(100, 105, 5, true),
+				mkExtentWithStepAndFormat(110, 115, 5, false),
+				mkExtentWithStepAndFormat(120, 125, 5, true),
+				mkExtentWithStepAndFormat(220, 225, 5, false),
+				mkExtentWithStepAndFormat(240, 250, 5, true),
+			},
+			expectedUpdatedCachedEntry: []tripperware.Extent{
+				mkExtentWithStep(0, 50, 5),
+				mkExtentWithStep(60, 65, 5),
+				mkExtentWithStep(100, 200, 5),
+				mkExtentWithStep(220, 225, 5),
+				mkExtentWithStep(240, 250, 5),
+			},
+		},
 	} {
 		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
@@ -1073,7 +1280,7 @@ func TestResultsCacheMaxFreshness(t *testing.T) {
 	for i, tc := range []struct {
 		fakeLimits       tripperware.Limits
 		Handler          tripperware.HandlerFunc
-		expectedResponse *tripperware.PrometheusResponse
+		expectedResponse tripperware.Response
 	}{
 		{
 			fakeLimits:       mockLimits{maxCacheFreshness: 5 * time.Second},
@@ -1271,6 +1478,46 @@ func TestResultsCacheShouldCacheFunc(t *testing.T) {
 			require.Equal(t, tc.expectedCall, calls)
 		})
 	}
+}
+
+func TestResultsCacheFillCompatibility(t *testing.T) {
+	var cfg ResultsCacheConfig
+	flagext.DefaultValues(&cfg)
+	c := cache.NewMockCache()
+	cfg.CacheConfig.Cache = c
+	rcm, _, err := NewResultsCacheMiddleware(
+		log.NewNopLogger(),
+		cfg,
+		constSplitter(day),
+		mockLimits{maxCacheFreshness: 10 * time.Minute},
+		PrometheusCodec,
+		PrometheusResponseExtractor{},
+		nil,
+		nil,
+	)
+	require.NoError(t, err)
+	rc := rcm.Wrap(tripperware.HandlerFunc(func(_ context.Context, req tripperware.Request) (tripperware.Response, error) {
+		return parsedResponse, nil
+	}))
+	ctx := user.InjectOrgID(context.Background(), "1")
+	_, err = rc.Do(ctx, parsedRequest)
+	require.NoError(t, err)
+
+	// Check cache and make sure we write response in old format even though the response is new format.
+	tenantIDs, err := tenant.TenantIDs(ctx)
+	require.NoError(t, err)
+	cacheKey := cache.HashKey(constSplitter(day).GenerateCacheKey(tenant.JoinTenantIDs(tenantIDs), parsedRequest))
+	found, bufs, _ := c.Fetch(ctx, []string{cacheKey})
+	require.Equal(t, []string{cacheKey}, found)
+	require.Len(t, bufs, 1)
+
+	var resp tripperware.CachedResponse
+	require.NoError(t, proto.Unmarshal(bufs[0], &resp))
+	require.Len(t, resp.Extents, 1)
+	expectedResp := convertFromTripperwarePrometheusResponse(parsedResponse)
+	any, err := types.MarshalAny(expectedResp)
+	require.NoError(t, err)
+	require.Equal(t, any, resp.Extents[0].Response)
 }
 
 func toMs(t time.Duration) int64 {